hardcoding cloud_850 to use publish-test key

+++ spec/destroy_ec2_ssh_keys.rb
@@ -13,16 +13,8 @@ ssh_dir = File.join(File.expand_path("~"), ".ssh")
 rest_yaml = File.join(File.expand_path("~"), ".rest_connection", "rest_api_config.yaml")
 rest_settings = YAML::load(IO.read(rest_yaml))
 
-if File.exists?(keys_file)
-  keys = JSON::parse(IO.read(keys_file))
-  keys.each { |cloud,hsh|
-    found = Ec2SshKeyInternal.find_by_cloud_id(cloud).select { |n| n.href =~ /#{hsh['ec2_ssh_key_href']}/ }.first
-    Ec2SshKey.new('href' => found.href).destroy if found
-  }
-  File.delete(keys_file)
-else
-  key_name = "#{ENV['RS_API_URL'].split("/").last}"
-  found = Ec2SshKeyInternal.find_by_cloud_id("#{cloud}").select { |obj| obj.aws_key_name =~ /#{key_name}/ }
-  found.each { |k| Ec2SshKey.new('href' => k.href).destroy }
-end
+key_name = "#{ENV['RS_API_URL'].split("/").last}"
+found = Ec2SshKeyInternal.find_by_cloud_id("#{cloud}").select { |obj| obj.aws_key_name =~ /#{key_name}/ }
+found.select { |k| k.aws_key_name =~ /monkey/ }.each { |k| Ec2SshKey.new('href' => k.href).destroy }
+File.delete(keys_file) if File.exists?(keys_file)
 rest_settings[:ssh_keys].each { |f| File.delete(f) if File.exists?(f) }
